Move Sep31DepositGeneratorSelf to the platform package.
