+++ platform/src/main/java/org/stellar/anchor/platform/SepConfig.java
@@ -27,8 +27,8 @@ import org.stellar.anchor.filter.Sep10TokenFilter;
 import org.stellar.anchor.horizon.Horizon;
 import org.stellar.anchor.paymentservice.circle.CirclePaymentService;
 import org.stellar.anchor.paymentservice.circle.config.CirclePaymentConfig;
-import org.stellar.anchor.platform.config.PropertyCircleConfig;
 import org.stellar.anchor.platform.data.*;
+import org.stellar.anchor.platform.service.Sep31DepositInfoGeneratorSelf;
 import org.stellar.anchor.sep1.ResourceReader;
 import org.stellar.anchor.sep1.Sep1Service;
 import org.stellar.anchor.sep10.JwtService;
@@ -37,7 +37,6 @@ import org.stellar.anchor.sep12.Sep12Service;
 import org.stellar.anchor.sep24.Sep24Service;
 import org.stellar.anchor.sep24.Sep24TransactionStore;
 import org.stellar.anchor.sep31.Sep31DepositInfoGenerator;
-import org.stellar.anchor.sep31.Sep31DepositInfoGeneratorSelf;
 import org.stellar.anchor.sep31.Sep31Service;
 import org.stellar.anchor.sep31.Sep31TransactionStore;
 import org.stellar.anchor.sep38.Sep38QuoteStore;
